refactor: using np.float64

+++ tra_go/core/win_graph.py
@@ -104,23 +104,23 @@ class WinGraph:
 
         self.is_model_worth_saving, self.is_model_worth_double_saving = simulation.get_is_worth_values()
 
-        fraction_valid_pred: float = np.mean(valid_pred.astype(np.float32))
+        fraction_valid_pred: float = np.mean(valid_pred.astype(np.float64))
 
-        fraction_max_inside: float = np.mean(max_inside.astype(np.float32))
+        fraction_max_inside: float = np.mean(max_inside.astype(np.float64))
 
-        fraction_min_inside: float = np.mean(min_inside.astype(np.float32))
+        fraction_min_inside: float = np.mean(min_inside.astype(np.float64))
 
-        fraction_average_in: float = np.mean(average_in.astype(np.float32))
+        fraction_average_in: float = np.mean(average_in.astype(np.float64))
 
-        fraction_win: float = np.mean(wins.astype(np.float32))
+        fraction_win: float = np.mean(wins.astype(np.float64))
 
-        all_days_pro_arr: NDArray = (max_pred / min_pred) * wins.astype(np.float32)
+        all_days_pro_arr: NDArray = (max_pred / min_pred) * wins.astype(np.float64)
         all_days_pro_arr_non_zero: NDArray = all_days_pro_arr[all_days_pro_arr != 0]
 
         all_days_pro_cummulative_val: float = np.prod(all_days_pro_arr_non_zero)
 
         pred_capture_arr: NDArray = (max_pred / min_pred - 1) * wins.astype(
-            np.float32,
+            np.float64,
         )
 
         total_capture_possible_arr: NDArray = max_true / min_true - 1
